always provide logging and exception for batch writes with retries
