+++ akka-projection-dynamodb/src/main/mima-filters/1.6.5.backwards.excludes/offset-store-dao.excludes
@@ -1,0 +1,2 @@
+# internal
+ProblemFilters.exclude[MissingClassProblem]("akka.projection.dynamodb.internal.OffsetStoreDao$BatchWriteFailed")
