+++ akka-projection-dynamodb/src/main/scala/akka/projection/dynamodb/scaladsl/Requests.scala
@@ -2,30 +2,22 @@
  * Copyright (C) 2024 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package akka.projection.dynamodb.javadsl
+package akka.projection.dynamodb.scaladsl
 
-import java.time.Duration
-import java.util.Optional
-import java.util.concurrent.CompletionStage
-import java.util.function.BiFunction
-import java.util.function.{ Function => JFunction }
-import java.util.{ List => JList }
-
-import scala.concurrent.ExecutionContext
+import scala.concurrent.Future
 import scala.concurrent.duration.FiniteDuration
-import scala.jdk.CollectionConverters._
-import scala.jdk.DurationConverters._
 import scala.jdk.FutureConverters._
-import scala.jdk.OptionConverters._
 
 import akka.actor.typed.ActorSystem
-import akka.japi.function.Procedure3
-import akka.projection.dynamodb.scaladsl
+import akka.projection.dynamodb.Requests.BatchWriteFailed
 import software.amazon.awssdk.services.dynamodb.DynamoDbAsyncClient
 import software.amazon.awssdk.services.dynamodb.model.BatchWriteItemRequest
 import software.amazon.awssdk.services.dynamodb.model.BatchWriteItemResponse
 
-object Retry {
+object Requests {
+  import akka.projection.dynamodb.Requests.batchUnprocessedTotal
+  import akka.projection.dynamodb.Requests.log
+  import akka.projection.dynamodb.Requests.retry
 
   /**
    * Perform a DynamoDB batch write, retrying writes for any unprocessed items (with exponential backoff).
@@ -36,35 +28,37 @@ object Retry {
    * @param minBackoff minimum duration to backoff between retries
    * @param maxBackoff maximum duration to backoff between retries
    * @param randomFactor adds jitter to the retry delay (use 0 for no jitter)
-   * @param onRetry called before each retry, with the response, the current retry count, and the delay for this retry
-   * @param failOnMaxRetries if max retries is reached, create a throwable for the failed future result
-   * @param system the actor system (for scheduling)
-   * @return all responses from attempts (in order)
+   * @return when successful return all responses from attempts,
+   *         otherwise [[akka.projection.dynamodb.Requests.BatchWriteFailed]] with last response
    */
-  def batchWrite(
+  def batchWriteWithRetries(
       client: DynamoDbAsyncClient,
       request: BatchWriteItemRequest,
       maxRetries: Int,
-      minBackoff: Duration,
-      maxBackoff: Duration,
-      randomFactor: Double,
-      onRetry: Procedure3[BatchWriteItemResponse, Integer, Duration],
-      failOnMaxRetries: JFunction[BatchWriteItemResponse, Throwable],
-      system: ActorSystem[_]): CompletionStage[JList[BatchWriteItemResponse]] =
+      minBackoff: FiniteDuration,
+      maxBackoff: FiniteDuration,
+      randomFactor: Double)(implicit system: ActorSystem[_]): Future[Seq[BatchWriteItemResponse]] =
     retryWithBackoff(
       request,
-      (request: BatchWriteItemRequest) => client.batchWriteItem(request),
+      (request: BatchWriteItemRequest) => client.batchWriteItem(request).asScala,
       (request: BatchWriteItemRequest, response: BatchWriteItemResponse) =>
         if (response.hasUnprocessedItems && !response.unprocessedItems.isEmpty)
-          Optional.of(request.toBuilder.requestItems(response.unprocessedItems).build())
-        else Optional.empty[BatchWriteItemRequest](),
+          Some(request.toBuilder.requestItems(response.unprocessedItems).build())
+        else None,
       maxRetries,
       minBackoff,
       maxBackoff,
       randomFactor,
-      onRetry,
-      failOnMaxRetries,
-      system)
+      onRetry = (response: BatchWriteItemResponse, retry: Int, delay: FiniteDuration) =>
+        if (log.isDebugEnabled) {
+          log.debug(
+            "Not all items in batch were written, [{}] unprocessed items, retrying in [{} ms], [{}/{}] retries",
+            batchUnprocessedTotal(response),
+            delay.toMillis,
+            retry,
+            maxRetries)
+        },
+      failOnMaxRetries = new BatchWriteFailed(_))
 
   /**
    * Retry generic requests with exponential backoff.
@@ -81,31 +75,17 @@ object Retry {
    * @param randomFactor adds jitter to the retry delay (use 0 for no jitter)
    * @param onRetry called before each retry, with the response, the current retry count, and the delay for this retry
    * @param failOnMaxRetries if max retries is reached, create a throwable for the failed future result
-   * @param system the actor system (for scheduling)
    * @return all responses from attempts (in order)
    */
   def retryWithBackoff[Request, Response](
       request: Request,
-      attempt: JFunction[Request, CompletionStage[Response]],
-      decideRetry: BiFunction[Request, Response, Optional[Request]],
+      attempt: Request => Future[Response],
+      decideRetry: (Request, Response) => Option[Request],
       maxRetries: Int,
-      minBackoff: Duration,
-      maxBackoff: Duration,
+      minBackoff: FiniteDuration,
+      maxBackoff: FiniteDuration,
       randomFactor: Double,
-      onRetry: Procedure3[Response, Integer, Duration],
-      failOnMaxRetries: JFunction[Response, Throwable],
-      system: ActorSystem[_]): CompletionStage[JList[Response]] =
-    scaladsl.Retry
-      .retryWithBackoff(
-        request,
-        (request: Request) => attempt.apply(request).asScala,
-        (request: Request, response: Response) => decideRetry.apply(request, response).toScala,
-        maxRetries,
-        minBackoff.toScala,
-        maxBackoff.toScala,
-        randomFactor,
-        (response: Response, retries: Int, delay: FiniteDuration) => onRetry.apply(response, retries, delay.toJava),
-        (response: Response) => failOnMaxRetries.apply(response))(system)
-      .map(_.asJava)(ExecutionContext.parasitic)
-      .asJava
+      onRetry: (Response, Int, FiniteDuration) => Unit,
+      failOnMaxRetries: Response => Throwable)(implicit system: ActorSystem[_]): Future[Seq[Response]] =
+    retry(request, attempt, decideRetry, maxRetries, minBackoff, maxBackoff, randomFactor, onRetry, failOnMaxRetries)
 }
